added cidr notation option to calculator, removed wildcard

+++ app/src/main/java/com/example/subnetter/model/NetworkInformation.kt
@@ -5,5 +5,5 @@ data class NetworkInformation(
     val broadcastAddress: IpAddress,
     val firstUsableAddress: IpAddress,
     val lastUsableAddress: IpAddress,
-    val numberOfHosts: Int
+    val numberOfHosts: Long
 )
